If-conversion optimization for Cminor
Add an experimental pass that converts some if/then/else statements
into "select" built-ins, which in turn can be compiled down to
branchless instruction sequences if the target architecture supports them.

The statements that are converted are of the form

    if (cond) { x = a1; } else { x = a2; }
    if (cond) { x = a1; }
    if (cond) { /*skip*/; } else { x = a2; }

where a1, a2 are "safe" expressions, containing no operations that can
fail at run-time, such as memory loads or integer divisions.

To support the optimization, I added a type inference and type checking
pass for Cminor, in backend/Cminortyping.v.  The inferred type information
is used to determine the type at which the "select" built-in operates,
and to prove semantic preservation.